requirements.txt
Django==1.10.3
gunicorn==19.4.5
virtualenv==13.1.2
djangorestframework==3.5.3
Pillow==3.1.0
django-filter==0.12.0
boto==2.39.0
filechunkio==1.6
django-cachalot==1.2.0
redis==2.10.5
python-memcached==1.57
django-redis==4.3.0
djangorestframework-xml==1.3.0
django-queryinspect==0.1.0
django-crispy-forms==1.6.0
django-rest-swagger==2.1.0
codecov
coverage
flake8
geopy
django-leaflet
